src/requirements-app.txt
fastapi==0.111.1
uvicorn==0.30.3
chromadb==0.5.5
mlflow==2.15.1
langchain-chroma==0.1.1
langchain-community==0.2.6
# langchain-core==0.2.10
langchain-openai==0.1.10
langchain==0.2.16 
langchain-core==0.2.41 
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
nemoguardrails==0.10.1
openpyxl==3.1.2
httpx==0.27.2